remove groupVar from rownames(lsm) carefully
groupVar string was removed from a contrast string using 'gsub' command with surprising outcomes in a few cases. For example, 'MUTATIONMUTATION - MUTATIONnormal' would be modified to ' - normal'. Seeing this in the results, the user could wonder if the correct set of samples had been used in the comparison. By limiting this gsub operation and performing it twice, we avoid this cosmetic problem.